% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{extract_univariate_params}
\alias{extract_univariate_params}
\title{Extract parameters from univariate Hawkes TMB object}
\usage{
extract_univariate_params(obj)
}
\arguments{
\item{obj}{TMB model object from fit_hawkes or fit_hawkes_multi_series}
}
\value{
List with mu, alpha, and beta parameter estimates
}
\description{
Extract parameters from univariate Hawkes TMB object
}
