% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_coefs}
\alias{get_coefs}
\title{Extract parameter estimates from TMB sdreport}
\usage{
get_coefs(obj)
}
\arguments{
\item{obj}{TMB model object with fitted parameters}
}
\value{
Matrix of parameter estimates and standard errors
}
\description{
Extract parameter estimates from TMB sdreport
}
