% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nhanes.r
\name{process_covar}
\alias{process_covar}
\title{Merge non-accelerometry data for NHANES waves G and H}
\usage{
process_covar(
  waves = c("G", "H"),
  varnames = c("SDDSRVYR", "WTMEC2YR", "WTINT2YR", "SDMVPSU", "SDMVSTRA", "RIDAGEMN",
    "RIDAGEEX", "RIDAGEYR", "RIDRETH1", "RIAGENDR", "BMXWT", "BMXHT", "BMXBMI",
    "DMDEDUC2", "ALQ101", "ALQ110", "ALQ120Q", "ALQ120U", "ALQ130", "SMQ020", "SMD030",
    "SMQ040", "MCQ220", "MCQ160F", "MCQ160B", "MCQ160C", "PFQ049", "PFQ054", "PFQ057",
    "PFQ059", "PFQ061B", "PFQ061C", "DIQ010", "LBXTC", "LBXHDD", "LBDHDD", "BPXSY1",
    "BPXSY2", "BPXSY3", "BPXSY4"),
  localpath = NULL,
  extractAll = FALSE,
  exclude_files = c()
)
}
\arguments{
\item{waves}{character vector with entries of (capitalized) letter of the alphabet
corresponding to the NHANES wave of interest. Defaults to a vector containing "G" and "H"
corresponding to the NHANES 2011-2012 and 2013-2014 waves.}

\item{varnames}{character vector indicating which column names are to be searched for.
Will check all .XPT files in located in the directory specified by localpath.
If extractAll = TRUE, then this argument is effectively ignored.}

\item{localpath}{file path where covariate data are saved. Covariate data must be in .XPT format.}

\item{extractAll}{logical argument indicating whether all columns of all .XPT files in the
search path should be returned. Defaults to FALSE.}

\item{exclude_files}{character vector of file patterns to exclude from processing.}
}
\value{
This function returns a list with elements equal to the number of waves specified.
Each element is named Covariate_* where * corresponds to the wave letter.
}
\description{
This function retrieves and merges covariate data from one or more NHANES data files
across waves G and H (2011-2014). Variables are merged using the NHANES unique subject
identifier (SEQN).
}
\details{
This function has been adapted for NHANES 2011-2014 waves (G and H). It searches all .XPT
files which match the NHANES naming convention and merges variables using SEQN.
}
